+++ src/trains-schedules/entities/trains-schedule.entity.ts
